+++ rubato/structure/gameobject/tilemap/__init__.py
@@ -1,0 +1,3 @@
+"""Tilemap components"""
+from .simple import SimpleTilemap
+from .tiled import Tilemap